Unite FunctionTestCase, LinkTestCase and related
